feed-grafana: check an environment variable before launching push scripts

+++ gsi/tr/feed_grafana.sh
@@ -3,7 +3,11 @@
 # This script is used to launch the push scripts that
 # send GMT monitoring data to Grafana running on the tsl019 host.
 
-GRAFANA_HOST=tsl019.acc.gsi.de
+# TOS_GRAFANA_HOST (domain name or IP address)
+if [ -z "$TOS_GRAFANA_HOST" ]; then
+    echo "Need to set TOS_GRAFANA_HOST. Exit!"
+    exit 1
+fi
 
 # Terminate running scripts
 killall push-dm-mon.sh
@@ -14,9 +18,9 @@ killall push-diag-mon.sh
 killall push-uni-mon.sh
 
 # Start scripts
-push-dm-mon.sh $GRAFANA_HOST &
-push-prod-lm-mon.sh $GRAFANA_HOST &
-push-gw-mon.sh $GRAFANA_HOST &
-push-tr-mon.sh $GRAFANA_HOST &
-push-diag-mon.sh $GRAFANA_HOST &
-push-uni-mon.sh $GRAFANA_HOST &
+push-dm-mon.sh $TOS_GRAFANA_HOST &
+push-prod-lm-mon.sh $TOS_GRAFANA_HOST &
+push-gw-mon.sh $TOS_GRAFANA_HOST &
+push-tr-mon.sh $TOS_GRAFANA_HOST &
+push-diag-mon.sh $TOS_GRAFANA_HOST &
+push-uni-mon.sh $TOS_GRAFANA_HOST &
